refactor(st-09082): centralize relational row extraction
