+++ packages/tools/src/data/relational/query/stream-executor-chunks.ts
@@ -4,7 +4,7 @@ import {
   resolveChunkSize,
   resolveTotalRowsLimit,
 } from './stream-executor-options.js';
-import { extractRows } from './stream-executor-runtime.js';
+import { extractRows } from './row-extraction.js';
 import type { SqlExecutor } from './types.js';
 import type {
   StreamingSelectChunk,
