+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-07-01
-**Active Story:** ST-09082 - Centralize Relational Row Extraction Helpers (Ready)
+**Last Updated:** 2026-07-03
+**Active Story:** ST-09082 - Centralize Relational Row Extraction Helpers (In Progress)
 
 ---
